chore: add preferences reducer
