+++ ts/features/common/store/reducers/index.ts
@@ -35,9 +35,7 @@ import {
   fastLoginReducer,
   FastLoginState
 } from "../../../fastLogin/store/reducers";
-import walletPersistor, {
-  WalletState
-} from "../../../newWallet/store/reducers";
+import walletReducer, { WalletState } from "../../../newWallet/store/reducers";
 
 type LoginFeaturesState = {
   testLogin: TestLoginState;
@@ -54,7 +52,7 @@ export type FeaturesState = {
   whatsNew: WhatsNewState & PersistPartial;
   loginFeatures: LoginFeaturesState;
   payments: PaymentsState;
-  wallet: WalletState & PersistPartial;
+  wallet: WalletState;
 };
 
 export type PersistedFeaturesState = FeaturesState & PersistPartial;
@@ -72,7 +70,7 @@ const rootReducer = combineReducers<FeaturesState, Action>({
     fastLogin: fastLoginReducer,
     cieLogin: cieLoginReducer
   }),
-  wallet: walletPersistor
+  wallet: walletReducer
 });
 
 const CURRENT_REDUX_FEATURES_STORE_VERSION = 1;
