+++ ts/features/newWallet/store/selectors/index.ts
@@ -9,5 +9,5 @@ export const isWalletPaymentsRedirectBannerVisibleSelector = (
   pipe(
     state,
     selectWalletFeature,
-    wallet => wallet.shouldShowPaymentsRedirectBanner
+    wallet => wallet.preferences.shouldShowPaymentsRedirectBanner
   );
